requirements.txt
crewai
crewai[tools]
langchain-google-genai
google-generativeai
fpdf2
requests
python-dotenv
google-adk
serpapi